requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
Django==5.1
django-cors-headers==4.9.0
django-split-settings==1.1.0
elastic-transport==9.2.1
elasticsearch==9.3.0
Faker==13.15.0
flake8==4.0.0
idna==3.11
mccabe==0.6.1
packaging==26.0
psycopg2-binary==2.9.11
pycodestyle==2.8.0
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
pyflakes==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uWSGI==2.0.31
wheel==0.46.3
